requirements.txt
torch>=2.0.1
torchvision>=0.15.2
numpy>=1.23.5
lightning==2.4.0
torchmetrics==1.6.1
timm>=0.9.16
argparse>=1.4.0
wandb>=0.15.8
pillow>=9.3.0
efficientnet-pytorch>=0.7.1
madgrad>=1.2
git+https://github.com/MIC-DKFZ/dynamic-network-architectures.git
rich>=12.6.0
matplotlib>=3.6.2
hydra-core>=1.3.2
batchgeneratorsv2>=0.2.3
blosc2>=3.0.0
peft>=0.14.0
git+https://github.com/MIC-DKFZ/acvl_utils.git
hd-bet
